backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.0
sqlalchemy==2.0.30
pydantic==2.7.1
httpx==0.27.0
gTTS==2.5.1
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
